Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

Temporarily removing zoom functions for Droid.  

When switching to Camcorder, the MediaRecorder immediately takes control of the camera, preventing any direct control of it via mCameraDevice.  The droid camera driver hangs randomly if setParameters is used to control zoom.  I think the only way to implement zoom on this device would be to plumb in smooth zoom calls in the MediaRecorder itself, in the framework native code.

I'll revisit it later if demand calls for it, but all other features appear to work in the Camcorder, making it still a highly desirable app to use - I didn't think this shortcoming should hold off progress in the short term.

Patch-set: 3