Update patch set 1
Patch Set 1: (1 inline comment)

Not ready to merge, mainly wanted to point out preferences for focus mode not being applied.

Rest of patch is to hide or enable the zoom control depending on whether or not we're in continuous mode (I have less control over the camera in camcorder mode so I can't workaround like in camera).

Problem is when I switch to a zoomable mode (infinity) from being in continuous, zoom freezes the app.  It doesn't if I start out in infinity to begin with.

Patch-set: 1