Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Not working, the framework is setting values like "touch-aec=-1x-1, touch-af-aec=touch-on, touch-focus=-1x-1" for anywhere touched. HTC's is of course douched up, and handles this weirdly- if I remember right, touch-aec is just "on/off" and touch-focus is x,y or -1,-1 to disable.

Patch-set: 1