Update patch set 1
Patch Set 1:

Do your modifications, then git commit <file> (or -a) --amend. That'll attach the modifications to your most recent commit instead of creating a new one.

You can even modify the title and description of the commit. As long as the "Change-Id" line in the description stays the same, you can re-run "repo upload", and it'll appear as a new patchset in this change.

Alternatively, you can begin from scratch and just paste the Change-Id line into a new commit's description, which will also become a new patchset for this change.

Patch-set: 1