Update patch set 2
Patch Set 2:

Preview size and actual video capture size are two different things. You shouldn't be querying the preview size to get capture capabilities of the device. The clause should be false. The reason it might not be is because you have the following line in setCameraParameters():
mParameters.setPreviewSize(mProfile.videoFrameWidth, mProfile.videoFrameHeight);

But that's a mistake, as you can see if you look at the logcat of that call you'll see the tegra 2 libs saying:
-----------------------------------------------
V/NvOmxCameraSettings( 1007): Changed: preview-size: 1600x1200 -> 1920x1088 {176x144,320x240,352x288,480x480,640x360,640x480,720x480,720x576,800x448,800x480,1280x720,1600x1200}
E/NvOmxCameraSettings( 1007): requested resolution 1920x1088 is unsupported, defaulting to 1600x1200
-------------------------------------------------
So, every time setCameraParameters() is called, the preview size is changed to 1600x1200 (see above), which prevents 1080p@30fps (no idea why).
So to keep 1080@fps30 enabled, you'd have to change it back to 1280x720 every time you've called setCameraParameters().

The expected behavior of the code is that first recording is 30fps, but any action causing a call to setCameraParameters() is not. Which is exactly what I get when I run the code.

Patch-set: 2