Update patch set 2
Patch Set 2:

Yes, the clause is true. But only because enableHighFrameRateFHD() happens to be called just after a call to setCameraParameters(). My point was that it's an unsupported setting and is only ever true of mParameters, never mCameraDevice.getParameters().

Any call to setCameraParameters() that does not go through startPreview() will cause 30fps to be disabled. At least, that's what I get.

Patch-set: 2