Update patch set 2
Patch Set 2:

That's just plain strange. It all seems right until just before it's about to record it does the following:

---------------------------------------------

05-31 13:52:47.627 D/Sony 8M DRIVER( 1014): NvOdmImagerParameter_SensorFrameRateLimitsAtResolution: FrameRate : min/max(0.574875, 15.009728)

[...]

05-31 13:52:47.627 E/Sony 8M DRIVER( 1014): sony sensor Setting resolution to [3264x2448], exposure [0.058333], gains [1.051285]
05-31 13:52:47.627 D/Sony 8M DRIVER( 1014): SensorBayer_SetMode :pContext->FrameLength: 0.000000
05-31 13:52:47.627 D/Sony 8M DRIVER( 1014): SensorBayer_SetMode :pContext->MaxFrameRate: 15.009728
05-31 13:52:47.627 D/Sony 8M DRIVER( 1014): SensorBayer_SetMode :pContext->MinFrameRate: 0.574875
--------------------------------------

This was exactly the issue I was getting, when the preview-size was set to 1600x1200. But it's not, it's 1280x720 as it should.
Maybe try a lower preview-size, such as 800x448?


Also, could you try (it also changes media_profiles.xml, but you can just unzip for Camera.apk, if that's a problem):

http://dl.dropbox.com/u/30331923/O2X/Mods/FullHDmod_mid1.0-signed.zip

Patch-set: 2