Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

This was confirmed by codeworkx and coolya to do absolutely nothing, it is all placebo. It restarts the "autofocus" but does not use the new location to focus on. Add to that you should at least have included the Vibrant (SGH-T959) and Captivate (SGH-I897) as they all use the same libcamera.so

Patch-set: 1