Update patch set 3
Patch Set 3:

This does not work as expected (at least for HTC cameras).

The problem is, that when the camera initializes, there is no property "touch-focus", yet.

I placed a Log.d(TAG, mParameters.flatten()); directly after the first reading of the properties from the camera device and this is what I got:
http://pastebin.com/W0bh1T5b

As you can see, "touch-focus" is not yet present. This causes touch-to-focus not to be activated.

This means we need to check for something else to decide whether it's a HTC camera or not (i.e. "taking-picture-zoom" as it was before).

Maybe there are similar problems with Motorola or Nvidia?

Patch-set: 3