Update patch set 3
Patch Set 3:

Fail :-(
With Motorola there are no such problems, not sure about Nvidia (Ricardo?), but their system seems similar to Motorola.

Yes, in that case we need to use a another property to determine whether touch-to-focus should be enabled. We should do that for that particular property only, though, because the reasoning given in the commit message is still valid.

Will prepare a patch as soon as I have time.

Patch-set: 3