Update patch set 1
Patch Set 1:

This sounds strange, but please try going to photo mode, change "Flicker reduction" to "50 Hz". CLOSE and RE-OPEN the camera application (the setting only seems to get applied this way). Switch to video mode and try again.

"50 Hz" is the mode that Samsung's application always uses and I was going to make a patch for it later. Didn't realize that it was a prerequisite for continuous autofocus!

Patch-set: 1