Update patch set 1
Patch Set 1:

Are you sure? It's in the photo mode and not the camcorder mode.

Anyway, it seems to work on-off for me on my I9000 as well. It stops working when I change my "flicker reduction" mode and restarts working after I close and open the camera application a few times.

Will do more research on this.

Patch-set: 1