Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

renamed string for continuous-video from "Auto" to "Auto Video" because it's only needed for videomode.

if we set continuous-video in photomode and switch to videomode, autofocus works like it should.

in videomode nearly all settings are unavailable for choosing.

Patch-set: 2