Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

Do the other focus modes work for video though (isn't select-able in stock Samsung)? I think it is better to leave it forced to continuous for now.

And I don't think continuous should be added as a select-able option because it makes no sense in photo mode.

As for touch-focus, I have a patch of it working (at least for aries), but it's pretty intrusive in the system (modifies the framework). I'm waiting until my continuous focus patch to be cleared before submitting it. (Because it depends on it.)

Patch-set: 2