Update patch set 3
Patch Set 3:

I would think that that belongs to the setContinuousAf function. And it would be better to check if continuous-video is a supported mode and set it to that regardless of the camera in question.

Can you check if the S2's continuous focus will work without continuous_af and mSamsungSpecialSettings as it seems like the S2 uses a different, more sane, property interface. Samsung's Camera.apk still sets all those (including continuous_af) for S2, but it may just be left-over code.

I will test and submit a new patch later when I have access to my computer.

Patch-set: 3