Update patch set 1
Patch Set 1:

To be honest, this looks pretty hackish to me. You're not informing AudioManager that recording is starting, you're setting some parameters highly specific to at least HTC, if not the ace, in generic code.

As I am not familiar with the audio code (yet) I don't have a good solution off-hand, hence no -1, but I think setting a generic parameter (e.g. "video_recording=on") and properly acting on that in libaudio seems like a better solution to me.

Patch-set: 1