Update patch set 10
Patch Set 10: Looks good to me, but someone else must approve

Working well on captivate. IMO, 40, 50 and 60 seconds are all longer than people would realistically use, and 5 seconds would be useful as an additional option. Also, as a "bell and whistle", i suggest some kind of audio warning that the pic is about to be snapped..... maybe a beep at 3, 2, 1 then click.

Patch-set: 10