Update patch set 1
Patch Set 1:

No, the s3d params are TI's, not LG's.

Also, that method has 2 purposes: setting the parameters, and adding the focus mode to the list (through the return value). We don't need the second on OMAP, if it is supported, it's already in the list.

Lastly: there are 2 reasons for the hackery: compatibility with possible future binary blobs on other devices, and doing this properly inside libcamera would still require a hook in the app to use the right coordinate param, so there's no gain in modifying the HAL to always export the param

Patch-set: 1