Update patch set 9
Patch Set 9:

Greg,

not exactly all of them ;) The static boolean caching the setting is still there.

I intended to fix that, but found a couple of things that I had issues with on the way, so now I'm sitting on quite a large diff which I can't test because my device doesn't have a switchable storage.

The diff is here: http://pastebin.com/0PPvFi8s

It fixes the following things:
- Get rid of the preference caching boolean in ImageManager.java
- Determine the availability of the option by the existance of a switchable volume pair. Previously, it was determined from the existance of a removable SD card, which is a necessary, but not sufficient condition. One doesn't want to see the option if there is nothing to switch inbetween.
- Allow the user to go back to following the volume switching for image storage. Previously, once the setting was touched once, there was no way to ever come back to a 'follow whatever is selected in CMParts' setting. Now, the setting toggles between 'use CMParts setting' and 'force external SD card'. If desired, this could also be made a ListPreference to also include 'force internal SD card'.
- Also determine the storage paths from the switchable pair itself instead of from various sources (env variable, Environment class)

Greg, can you please give this patch a try and include it if it works fine?

Patch-set: 9