Update patch set 1
Patch Set 1:

@Yanuar: In your case you should set ro.device.screenrotation=270 in your build.prop instead of adding +90 in the code. That is the sole and only purpose of this change!

As you can read, mDeviceScreenRotation is ro.device.screenrotation, so 0+90 = -270 (%360) which is the value you seem to need.

Regardind CMParts, that property isn't thought to be changed and or tweaked by the user. It's linked to the device, in order to correct a wrong rotation, not a setting for the user. I hardly think of people wanting to have their picture rotated. 

Imho I think you are misunderstanding the purpose of this change.

Patch-set: 1