Update patch set 2
Patch Set 2:

I'm not seeing a summary for this setting (I apologize if I missed it). I would recommend be added (if one is not there already).

A possible summary would be something along these lines, "Allow the usage of the Power button to take pictures or start & stop recording of videos".

Patch-set: 2