Update patch set 1
Patch Set 1: No score

(1 inline comment)

It has tap to focus, but it is buggy. It doesn't switch into Macro mode for instance. People also prefer to have manual control over it.

Anyway, tap-to-focus mode is still enabled (it adjusts based on the focus mode set in the settings, so if you're in Macro mode, tapping the screen will focus only in the macro range).

Patch-set: 1