Update patch set 2
Patch Set 2:

Just to clarify the logic, right now this is a software toggle to allow testing by simply relaunching the application when pushing test camera files. If a resolution is not found and this becomes a permanent fix, it will be reduced down to just the xml / string portion.

Patch-set: 2