Update patch set 1
Patch Set 1:

I should note that when other picture formats are made available, this code can be updated, but currently those are the only two offered by the frameworks and the code is flexible enough to allow transition to others should they be added. This is hardly a reason to say the code is wrong based on other available formats, though, when there is no support for them in frameworks anyway.

Patch-set: 1