Update patch set 1
Patch Set 1:

And your point is...?
The framework changes add the ability to set/get picture formats, they don't use them, and it has no impact on other platforms. They're methods, not applications of methods.

This change is an application of those changes without any further verification, and it will break anything that isn't a qualcomm. CAF is qualcomm, CM isn't

Patch-set: 1