Update patch set 1
Patch Set 1:

As I stated before it was a direct port of the feature, not a necessary feature. It can always be restored should spending the time required become an option. Often times there are a few here with a very specific, previously established layout in mind so this way they will have an open floor to submit those ideas instead of trying to direct another change toward them.

Patch-set: 1