Update patch set 1
Patch Set 1:

I'm on the fence about this.  On my DSLR, the icon only shows up in "Off" and "Fill-flash" modes.  On the automatic flash mode, it doesn't show anything on the UI at all.

Given that auto is the default state for devices and people usually assume the camera application to be in auto mode, it may be a good idea to have it on at all times.

Nebkat: why don't you want to the icon to be there?  It's not really distracting.  The only issue I see is that it breaks the consistency in the Camera UI (GPS icon is only visible when geolocation is on, but that feature isn't as widely used)

Patch-set: 1