Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I would equate a smartphone to a point and shoot camera, and in my experience most P&S have the flash indicator showing all the time (well at least when she anything else on the screen). And since smartphones are typically just as a P&S, most people will not want to fiddle with the settings too much, they just want to *point* and *shoot*. Having the flash set to On, Off or Auto can dramatically alter the out come of the picture and just displaying the indicator all the time would make pointing and shooting easier (at least seeing if the flash is on) for the end-user. Now if smartphone cameras were on par with DSLR's this may be a different situation but they aren't and this is why I'm +1ing this.

Patch-set: 1