Update patch set 1
Patch Set 1:

Quite a few tabs in here - the standard in the code is spaces - not tabs for indentation.

Also the conversion of the preference of off to a boolean could be made more elegant by storing it as a boolean rather than a string - however I don't think that matters enough.

Code all seems good - however as Daniel points out the logic between SDCard and internal RAM is swapped.

Finally - one of the main features of ICS was to use RAM to store photos, before sending them to the persistent storage - to improve photo taking speed - I assume this change has no effect on that?

Patch-set: 1