Update patch set 1
Patch Set 1:

Jesus,
please convert tabs to spaces and remove trailing spaces in the source code. I tested this on i777, works well. I see that directories for internal and external memory are hard-coded. In CM7 system settings there is an option to reverse them. I'm not sure if such an option will be added to CM9 - in that case your code would need to take this into account. See this change for CM7 Camera I made to take this into account:

http://review.cyanogenmod.com/#/c/10840/

For now it works well though, at least on Galaxy S2 phones, so I'd like to see it approved. Should be tested though on other phones first as well.

Patch-set: 1