Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

In my opinion, to truly fix the issue on the camera app, we should drop the inline selector thing altogether, and replace it with a button which opens an independent settings panel, kind of similar to the flow you see when on Settings -> Storage -> Menu -> USB Connection. On phones with a single storage option, the button would be hidden.

These series of patches use abbreviations which might not be familiar to users and might not reflect the actual phone reality; they also require changes on several parts of the android system. On the other hand, adding a settings panel would only require new code on the camera application.

Patch-set: 1