Update patch set 3
Patch Set 3:

Your objection has been noted, but you are simply restating the obvious. It has been clarified through previous posts that this is NOT a final submission meant for merge and only a reference during HAL testing. You are in no way required to use this in your local source.

Patch-set: 3