Update patch set 3
Patch Set 3:

No worries, I just didn't want it confused as being a requirement for anything. This is just a quick, sloppy way to get a preview until something is finalized. The string is used to allow editing from within the build to be a little easier so the app toggle switch could be removed.

Patch-set: 3