Update patch set 3
Patch Set 3:

Reason i did so, was to keep it backward compatible with those who had it set to off. However seeing as the popup menu itself crashes already, seems a wipe of the camera preferences will be required anyway, and so no need to set a default in the catch prefix. I'll put in a warning in the commit msg.

Patch-set: 3