Update patch set 4
Patch Set 4:

You have a default value:
<string name="pref_camera_timer_default" translatable="false">0</string>

which you set it up in preference.xml, but not in the code. So if you decide latter, that it will be good to change default value to 20, you also will need to change this in the code.

So the better way is to set it as the default value in the code too.

See the code above you set it manually as "0", but should be getString(R.string.pref_camera_timer_default)

Patch-set: 4