Update patch set 1
Patch Set 1: No score

There seems to be an issue with settings. If I open settings, I see "save location" on top. If I scroll to the bottom, I see the same option again. But if I scroll up-and-down several times I see different settings on the top/bottom of the list, and often twice the same option. After removing both your camera patches, the problem is gone.

Patch-set: 1