Update patch set 4
Patch Set 4:

Burst function works on i9100g, but the camera preview doesn't reset/change between bursts. For example, if you change the camera's position while in the middle of burst mode, it will continue to show the camera preview of when you started the burst mode and it will not show the preview at the new position.

Patch-set: 4