Update patch set 4
Patch Set 4:

Burst works fine on both hercules and gnex, however on gnex if I use burst mode (tested 5 shot), click the preview, then click the picture to take me to gallery and press back to go back to camera it results in a black screen. Doesn't happen on hercules, only on gnex in my testing.

Also, should the arrow/options be reversed so it points to the right so it's inline with the other options present?

Patch-set: 4