Update patch set 1
Patch Set 1:

I am starting to think this was not such a good choice after all. The camera app also locks expsure settings together with focus wich can lead to an annoing situation when you are waitng for focus lock to drop to be able to see anything on the screen. Also I have now realised it is possible to lock the focus by pressing and holding the camera button down while reframing. I had not noticed this before as it only works when focusing with the camera button and not tuch to focus. A better solution would be to move back to 3 seconds and make holding the camera button locking it until release.

Patch-set: 1