Update patch set 2
Patch Set 2:

i think at the moment this patch only is useful to tegra3 camera on htc one x endeavoru. it resolves important issues. It's an optional overlay setting, so if a device tree needs it, it's there. Maybe the name of the setting (needsHTcCamMode) is too wide, something more specific would suggest the function of it better.

Patch-set: 2