Update patch set 1
Patch Set 1:

everything works as it should, however if you press the home key to leave the camera app and then re-open the camera app, Gallery FCs. It does this on both d2tmo and hercules. I noticed though that if I pressed the back key to leave the camera app it wouldn't FC when I re-opened it (tested on d2tmo), only when leaving it via the home button.

logcat - http://pastebin.com/YHT9A02s

Patch-set: 1