Update patch set 1
Patch Set 1: Do not submit

This is broken in so many ways :( 

Does it work? Did you author this code? What is it actually doing? It at least breaks every device in CM, that's for sure :) Why is this (brokenly) targeted at ARMv6, isn't this just an Adreno issue?

Patch-set: 1