Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

I'm sorry Rashed but for me it doesn't build. It fails on PanoramaActivity.java saying that mPausing and mMosaicView doesn't exist. 

I think you need to change it to mPaused and super. That's what I did and it worked for me.

Patch-set: 4