Update patch set 3
Patch Set 3:

Eddy Witkamp: It works like it should at least on maguro and several different devices tested before submitting this. The pinch area is only the preview area, not where the buttons are so sliding still shows the gallery. If it doesn't work for you, let us know what device and build you are running.

Patch-set: 3