Update patch set 1
Patch Set 1:

@Björn: The button-area will only swipe if you hit the grey area between the buttons. Everything else is hit and miss.
If you touch either the shutter button, the option buttons on the left or the last photo taken on the right, it will NOT swipe to the gallery.

On a SGS2 this only means you have about 2x 1 cm to make it swipe, it is totally useless this way in my opinion.

Patch-set: 1