Update patch set 1
Patch Set 1:

On a maguro and hated the behavior.  Expected to be able to swipe to the gallery and learned that I had to do it down below, also hate it because it makes it easier to hit one of the soft buttons by mistake.  I don't see the utility of pinch zoom in the camera at all, especially when there is a slider there already.

Patch-set: 1