Create patch set 6
Patch-set: 6
Subject: Camera: Add volume key to control the zoom in/out. I will make my phone in lanscape when I take picture or capture video. I feel it very not easy to adjust the picture zoom size by touch the screen, because the take picture button in the right side. The finger in the left hand is easy to push the volume key, so I make the volume key up/down as the zoom out/in.
Commit: 4d1c15adfeabc71618f2aded71c5847b936d6700