Update patch set 2
Patch Set 2:

Tested on I9100G. One thing that I have noticed is that, say you switch from Internal storage to external and you take pictures, when you swipe to the left to go to the gallery view, it won't show any of the pictures you just took as it'll still be looking at the internal storage folder rather than the external storage folder. This is fixed once you restart the camera. Otherwise, it works as described.

Patch-set: 2