Update patch set 2
Patch Set 2: No score

There's no way to obtain the minimum width/height from layout definition in xml. The API for that has been introduced in JellyBean.
The controls in Camera app are not displayed, the app is not usable. You cannot change settings, switch mode or take a picture. If you know a better way how to fix it please let me know.

Patch-set: 2