Update patch set 2
Patch Set 2: No score

But I know that certain drivers have another problem with this. The main issue is that "keep" feature which will not "release" the camera when switching between video and picture mode.

Certain drivers need to be reset completely when switching between video and picture mode or they will get screwed up.

Patch-set: 2