Update patch set 6
Patch Set 6:

This is still using a static modifiable variable.  This makes the code brittle as it's introducing state into the classloader, and has much less transparency for debugging if the value is set or not.

If no one else has an issue with this code, then I'm fine, but  I would take a look at passing the path as an argument for methods that need it, like so:
Storage.generateBucketIdInt(pathToStorage);

If anything, please check for the case where the storage string is null (not set yet) when the methods like Storage.generateBucketIdInt() are called.

Patch-set: 6