Update patch set 6
Patch Set 6: No score

I can't see another way this can be done? sStorage contains the path to where the pictures/videos will be stored. It needs to be updated (which is why it was public) when you select what storage you want to use. From this it needs to be shared between multiple classes that need the storage location.

This wouldn't be appropriate

Storage.generateBucketIdInt(pathToStorage);

since the pathToStorage is Storage.sStorage.

The AOSP source was designed to load the storage path into a 'public static final' variable since it never needed to change after it had been set. But now it does and I can't think of a better way to do this?

Patch-set: 6