Update patch set 6
Patch Set 6:

Yes, so now that it needs to be modifiable, it really shouldn't be a static variable (and this should be removed).  You're storing it in the preferences, which is good, and now you need to pipe that variable into the classes where it is used.

Usually the best way to do this is with dependency injection, where you construct the classes with the storage location as a constructor argument.  Since this is probably modifyable during the lifetime of the util classes and such, you'll want either:
* store the path in the classes that need it, and have setter methods for updating it on path change
* or include the path as an argument when you need it

Since the Storage class is a static util class, it would make sense to accept the path as an argument to those util methods, like my example before.

On the other hand, the activity base is used as an object, so you can have a setter on there.  If you go this route, set the storage path from the preferences (or as default) on creation, and then update it when the settings are changed.

You can also add the storage path as an argument on the ActivityBase methods that need it, so it would be something like this:
String storagePath = CameraSettings.readStorage(mPreferences);
mActivity.reuseCameraScreenNail(!mIsImageCaptureIntent, storagePath);


Basically, you should be removing that static variable, and instead either pass it as a function argument or keep it as a class variable.

Patch-set: 6