Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

I'll fix typos in next upload. I lost my old committer name because of change to .org domain and a stupid try to repair. So I've got new name "bla" and can't change it back to thunderger. :D

Please decide if a '-' should be used or not. Some translation make use of it other don't.

Patch-set: 2