Update patch set 3
Patch Set 3:

Probably xplodwild just didnt had the time to pick this back up again.

Last time we discussed, would be good to apply native code, as described in to do list and also renderscript so we can get much faster rendering. After that will be possible to try to implement a bunch of HDR math methods and an algorithm to select which on gives us better dynamic range. But we would really need much more processing speed for this.

Patch-set: 3