Update patch set 3
Patch Set 3:

Giacomo-
Don't get too excited. This was submitted during X-mas time, and has not seen a single revision since. It has potential, but no matter how it works from one device to another, the bigger problem is that it's unmergeable in it's current state,(no menu option, among larger stylistic issues) and it's likely been abandoned. So unless somebody wants to "adopt a patch", this one is dead in the water.

Patch-set: 3