Update patch set 2
Patch Set 2:

So to understand this correctly, there's a need for a config.xml value here so it can be coded depending on the need of the given device's camera driver?
I need this for htc one x endeavoru too, but we need a unified solution so everyone's still happy.
Maybe just add a boolean switch that defaults to the 'bad' handling and the exceptional devices like endeavoru can use the 'normal' handling.

Patch-set: 2