Update patch set 2
Patch Set 2:

Logic in the patch linked above for resolution changed. You should override the value only for Legacy devices that doesn't support the snapshot while video not paused. So for those devices that support it, you just need to apply the other patch linked above, and try. If you're on a legacy device, override the value to 'false' so it won't take snapshots while video is recording.

Patch-set: 2