Update patch set 5
Patch Set 5: I would prefer that you didn't submit this

This doesn't make sense to me. Why wouldn't you want to change the isVideoSnaphsotSupported directly? Is there a case where you want this function to return true while the new function does not? 

I'm on my phone and Gerrit is a bit weird to use so maybe I overlooked something. 

My approach would be to change the is video snapshot supported value at the lowest level possible. As far as I understand the code (and I don't really do) the value defaults to true but can be set by the camera hal to false. So maybe let the value default to false and hope that all camera hals report the correct value or make something configurable on a lower level. 

I will have a closer look again when I'm home. I think the current change is inconsistent if other parts trust the isVideoSnaphsotSupported value.

Patch-set: 5