Update patch set 5
Patch Set 5:

I see your point Thomas, but for some reason on some devices the default AOSP behavior seemed to cause trouble (maybe its a vendor blob problem there), as john c craig reported in the other patch. So I thought we could add a way to override the default behavior to what it has been changed already (mPaused -> !mPaused) but this time with this flag.

I'd be totally fine if the other formerly submitted patch could be merged, as that seems to be the last patch that endeavoru cm-10.1 would need for becoming a complete tree that could be built in nightly - jenkins.

Maybe this patch is superfluous, but I tried to add a solution that can suffice for all of the devices, new or legacy, and thus endeavoru can progress further as well.

Patch-set: 5