Update patch set 5
Patch Set 5:

@Nagaprasad, okay thanks, please do so.
This is a weird patch indeed, not easy to follow its logic, as Thomas Wendt stated. So please try this. :-)

Anyway, if we find out what's the real problem with john c craigs device then we can progress with the other patch which is more logical, close-to-AOSP patch http://review.cyanogenmod.org/#/c/29039/2
But this or the other patch, or some new patch needs to be merged to let the endeavoru and other new devices take snapshots when video is being recorded.

Patch-set: 5