Update patch set 6
Patch Set 6:

The other patch simply enables video snapshot for supported devices (which was accidentally disabled by my patch earlier) and should not affect taking picture at all. If there it a bug, it would be elsewhere and we should ask John to find and fix it in a separate patch set.

The way I see it, this patch is just an ugly and roundabout way to disable support for video snapshot for devices that do not properly support it. Checking if an activity is paused or not should not be device-dependent at all, no matter what it is.

Patch-set: 6