Update patch set 8
Patch Set 8:

Nice, the only better technique I can think of to improve performance is to do it in a place where you can operate directly on the image buffers without having to memcopy to userspace. You'd have to either get into the userspace HAL implementation (usually closed source), or do it in the kernel which would be difficult to port.

Patch-set: 8