Update patch set 16
Patch Set 16:

It works fine on Endeavour, but the animation of the camera app when taking a pic slows it down so much, that i cant hold the camera still long enough to take a HDR pic that isnt washed out.
How can one disable the camera/gallery2 animation for this to work better?

Patch-set: 16