Update patch set 3
Patch Set 3: Verified; Looks good to me, approved

Check out Filterscript too, it's a subset of RS that Google created for camera filters. It has the potential to perform a bit better. The documentation is basically non-existent, though.

Probably the last major missing piece (besides the algorithm itself) is to use burst mode on cameras that can support it. D2 lets you set the number of shots per snap with a parameter. We have support for burst shots in the app already.

Awesome :)

Patch-set: 3