Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

Do you really think that JPEG quality should be set to "Super" by default? 
On one hand I agree, because I keep my JPEG quality set as super, and it makes a fairly appreciable difference; but on the other hand, don't you agree that defaulting to the highest quality setting might bring headaches for some users with certain storage limitations? Suppose that someone with a storage limited device all of a sudden has this setting as their default, yet they're not aware because they didn't notice it on the change-log or whatever. Let's also assume they're flashing from stock. They'll probably expect the camera to take the same size photos as it did on stock, and this could cause their storage usage to be way above what they think it to be. I did a test on my Galaxy Nexus to find the average file size difference between "high" and "super" JPEG quality, and the average was a 33% increase in the file size. Some were increased as much as 46% in size, depending on what was being photographed. Anyway, I'm just thinking that this wouldn't be a good default setting for all devices across the board. Super quality JPEG's should be a setting that the user specifically chooses, otherwise they'll be unaware of the extra resources being used. Even if storage space isn't a concern for someone, the file size increase is still way above normal,which makes "super" a bad choice as a default. So, that's how I'm seeing it at the moment, but what do you think?

Patch-set: 3