Update patch set 3
Patch Set 3: No score

I'm waffling on this.  The continuous-video focus mode is in Android's official camera.parameters ( http://developer.android.com/reference/android/hardware/Camera.Parameters.html ), so that's a good inclusion.  On the other hand, hyperfocal focus and all of the scene modes are HTC specific and likely won't work with other manufacturers' camera props.

My opinion would be to cut this patch down to just the continuous-video focus mode and add the other modes as device overlays.

I'm definitely not the authority on this subject though...

Patch-set: 3