Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I think that this is not the default behavior for a reason. Many people, including myself, do not like unexpected rotations. When someone specifically disables auto-rotate as a system-wide policy, that means that they don't want anything rotating unless they activate it manually. The rotation animation is a little slow, as well as annoying if accidentally triggered, which could really get in the way of someone that is taking lots of consecutive photos.
 Honestly, I think the rotation setting should be enforced throughout the system, with very few exceptions. To me, it seems like a bad idea for the camera to be one of those exceptions. It works better in just the Gallery.
 Whatever happens, this should probably be made a user-setting if anything. Changing default behaviors like this one will likely just cause headaches later.

Patch-set: 1