Update patch set 1
Patch Set 1:

Why on earth wouldn't you want Camera to rotate? I personally keep autorotate off all the time, but when I'm in camera I do NOT want portrait pictures to be taken when I'm holding it in landscape mode. That is just annoying. I'm sorry but I don't understand at all. If anything, Gallery is the disputable one. I don't see any reason for Camera not to autorotate. Before 4.2 the individual UI elements would rotate regardless of autorotate setting, but this doesn't happen anymore which is why I submitted this patch.

Patch-set: 1