Update patch set 1
Patch Set 1: (1 inline comment)

I too am happy that there is someone who is the 'Translation lead'. I just would have never thought that would be me :)

As I need to stay in touch with the crew, I have heard their concerns and also the concerns of us, the translators. When a 'minor' update is released in AOSP, such as Android 4.2.2 for cm-10.1, there are many merge conflicts, because CyanogenMod intentionally modifies the source. The main problem behinds those merge conflicts are actually the translations, because:

- modified AOSP strings (by CM) have been updated in AOSP, but if they are not the same, the merge will fail;
- sometimes CyanogenMod additions are placed between AOSP translations, making the line numbers not consistent with AOSP's and the merge will fail.

The first one is something that's hard for us. I have also changed a lot of AOSP strings since there are inconsistencies lying around and sometimes a translation could be shorter or improved in any other way. Because of the merge conflicts, changes to AOSP strings will be blocked, unless the translations are _really_ wrong. Inconsistencies is not something that we want to have, but 'camera' and 'photo device' mean the same and that is why I am blocking that. There is not any difference between '...' and '\u2026' for the user and that is why I am blocking that.

The second one is easy to solve: resources should be added under AOSP strings, divided by a line containing '<!-- CYANOGENMOD ADDITIONS -->'. In the next version of Android (5.0?), the resources will be added to seperate files (cm_strings.xml, cm_arrays.xml and cm_plurals.xml).

Regarding your statement about the impact consistencies will have on the user: remember that these strings come from AOSP, so AOSP would be found amateurish in that case :)

Of course this discussion is still 'hot'. Feel free to respond to Jacob Müller's post in the Google+ Community.

Patch-set: 1